class.DwooTemplateHandler.php -- corrected path to match up with new folder layout class.navigation.php -- fixed settings issue with public losses branches/dev/img -- removed/moved branches/inc/common_paths.inc updated include path for new folder layout branches/dev/js -- removed/moved branches/dev/tpl -- removed/moved
branches/dev/style/karambitks/tpl <- new dir structure .. All related fields for the style/templates/css/js would be under /style Should be fairly self-explanatory.  stuff that is globally shared like css /images are in the root of the folder.. Stuff that are style specific such as layout or css will be in their own folders.